Setup for fixing rotations.
QWEASD rotate blocks. By selecting a block we only rotate blocks of
identical rotation. Unfortunately, this seems unsolvable. There are
groups of blocks that cannot be correctly oriented (as far as I can
see). Further investigation is necessary, perhaps the orientation is
context sensitive? This would be madness but this is FtD we are talking
about.